sched/fair: let scheduler skip util checking if cpu is idle
Current cpu util includes util of runnable tasks plus the recent
utilization of currently non-runnable tasks, so it may return a non-zero
value even there is no task running on a cpu. When scheduler is selecting
a cpu for a task, it will check if cpu util is over its capacity, what
could happen is that it will skip a cpu even it is idle, so let scheduler
skip util checking if the task perfers idle cpu and the cpu is idle.

Bug: 133284637
Bug: 144961676
Test: cpu selected as expected
Change-Id: I2c15d6b79b1cc83c72e84add70962a8e74c178b8
Signed-off-by: Rick Yiu <rickyiu@google.com>
Signed-off-by: Jimmy Shiu <jimmyshiu@google.com>

[hridaya: Fix conflict introduced by backport commit 444683e9df585
	("BACKPORT: sched/fair: Fix overutilized update in enqueue_task_fair()")]

Change-Id: I93fb8186c951e9604b9c6cdf4670c57ffee3df13
Signed-off-by: hridaya <info.hridayaprajapati@gmail.com>
Signed-off-by: NotZeetaa <rodrigo2005contente@gmail.com>