devfreq: memlat: Add CPU Hotplug and Idle awareness in memlat
Register with CPU Hotplug and Idle notifiers so that memlat
gets notified when CPU is being hotplugged in/out or when CPU
enters Idle state. The perf events should be recreated when the
cpu is brought back online. Also do not update the counts when
the cpu is hotplugged out or when it enters Idle state.

Change-Id: I35a05be0a43a4519282d5787e6c87d92e5bec1de
Signed-off-by: Rama Aparna Mallavarapu <aparnam@codeaurora.org>
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: NotZeetaa <rodrigo2005contente@gmail.com>