arm64: csum: Fix pathological zero-length calls.
In validating the checksumming results of the new routine, I sadly
neglected to test its not-checksumming results. Thus it slipped through
that the one case where @buff is already dword-aligned and @len = 0
manages to defeat the tail-masking logic and behave as if @len = 8.
For a zero length it doesn't make much sense to deference @buff anyway,
so just add an early return (which has essentially zero impact on
performance).

Change-Id: I02c9d6ae710981e1dc4b92a97aced000eb025b11
Signed-off-by: NotZeetaa <rodrigo2005contente@gmail.com>