sched/core: Don't force immediate migration in migrate_enable()
Constant forced migrations from migrate_enable() introduce a lot of jitter,
when it's not a big deal to just wait for the task migrate naturally the
next time it hits the scheduler. As such, don't bother enforcing a task's
allowed CPUs after migration is re-enabled. It won't take long for the task
to reach the scheduler again thanks to forced preemption, so the amount of
time the task can spend on the wrong CPU is short and bounded.

Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: NotZeetaa <rodrigo2005contente@gmail.com>