usb: gadget: f_fs: Remove unused variables
With commit [1], our original intention was to just remove
ipc log string and resolve the uninitialized variable warn
for our benefits. It seems commit [1] made that variable's
use case unused and also the ones who are dependent on it
and the similars.

This removes all of the unused variables detected by
GCC after commit [1].

[1]: 7f4960191eba2aee302cf9a731dc83b9256d01b9 {"usb: gadget: f_fs: Make ipc log strings no-ops"}

Change-Id: Ia021bd885c02e1a99a2f9fb6346ea2dfb96f24bb
Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: hridaya <hridaya@pixelos.net>
Signed-off-by: NotZeetaa <rodrigo2005contente@gmail.com>