msm: camera: isp: Handle RUP in applied substate
In RDI only context, if RUP is received earlier than SOF,
the state machine is set to applied state and there is no
handling for RUP in applied substate. Hence leading to
bubble scenario and apply failures.

To overcome such scenario, handling for RUP in applied
substate is introduced in RDI only context statemachine.
In this handling, substate is moved to EPOCH so when
SOF is scheduled next state machine moves in right
direction.

CRs-Fixed: 3371187
Change-Id: I44442a8aa85a996c0adc658be52cfe92b7ecfc4d
Signed-off-by: Shivakumar Malke <quic_smalke@quicinc.com>
Signed-off-by: NotZeetaa <rodrigo2005contente@gmail.com>