binder: Compile out binder_[return|command]_strings on !FTRACE
binder_return_strings and binder_command_strings arrays are only
needed when CONFIG_FTRACE is enabled. Since we are not using
CONFIG_FTRACE at the moment, we can compile these pieces of codes
out. To do that, just simply guard the arrays with CONFIG_FTRACE
so that compilation doesn't error out in the future if we ever
turn on CONFIG_FTRACE again.

Signed-off-by: Tashfin Shakeer Rhythm <tashfinshakeerrhythm@gmail.com>
Signed-off-by: NotZeetaa <rodrigo2005contente@gmail.com>